fixed parse_args struct scan * bug
need to use pointer to a pointer in this circumstance for the object to
remain in memory beyond the end of the function parse_arguments.

This completes parse _arguments() and friends, they're  now thoroughly
tested.

Also added a check in tcp's partial_handshake() to make sure the packet
we've just received isn't somehow bigger than IP_MAXPACKET, which is
the size of the buffer.